Removed all 65 code smells in popupmenu

+++ src/main/java/deco2800/skyfall/gamemenu/popupmenu/BlueprintShopTable.java
@@ -79,16 +79,9 @@ public class BlueprintShopTable extends AbstractPopUpElement {
         infoBar.setSize(650, 55);
         infoBar.setPosition(130, 435);
 
-//        Table infoPanel = new Table();
-//        infoPanel.setSize(410, 400);
-//        infoPanel.setPosition(25, 18);
-//        infoPanel.setBackground(generateTextureRegionDrawableObject("info_panel"));
-
         this.blueprintPanel = new Table();
-        // updateChestPanel(chest);
 
         blueprintTable.addActor(infoBar);
-        //blueprintTable.addActor(infoPanel);
         blueprintTable.addActor(this.blueprintPanel);
         blueprintTable.setVisible(false);
         stage.addActor(blueprintTable);
@@ -115,15 +108,15 @@ public class BlueprintShopTable extends AbstractPopUpElement {
             ImageButton icon = new ImageButton(gameMenuManager.generateTextureRegionDrawableObject(b.getName()));
             icon.setName("icon");
             icon.setSize(100, 100);
-            icon.setPosition(xpos + (float) count * 130, ypos);
+            icon.setPosition(xpos + count * 130, ypos);
             if (isBought(b)) {
                 Label cost = new Label("X", skin, "white-label");
                 cost.setName(b.getName());
-                cost.setPosition(xpos + 85 + (float) count * 130, (float) ypos + 75);
+                cost.setPosition(xpos + 85 + count * 130, ypos + 75);
                 blueprintPanel.addActor(cost);
             } else {
                 Label cost = new Label("$" + b.getCost(), skin, "white-label");
-                cost.setPosition(xpos + 85 + (float) count * 130, (float) ypos + 75);
+                cost.setPosition(xpos + 85 + count * 130, ypos + 75);
                 cost.setName(b.getName());
                 blueprintPanel.addActor(cost);
